backend/requirements/base.txt
asgiref==3.2.7
Django==3.0.7
django-probes==1.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
djangorestframework==3.11.0
django-cors-headers==3.3.0
Collectfast==2.1.0
django-environ==0.4.5
argon2-cffi==20.1.0
django-allauth==0.42.0
django-rest-auth==0.9.5
Pillow==2.2.1